Proper SSA with graph representation in-memory
Previously the expressions were a bit too Scala-like
and were tricky to reason about in combination with SSA
concepts like phi nodes.

Now the situation is much closer to llvm's representation
with basic blocks that end with terminator instruction
that either transfers control flow to another basic block
or ends execution of the current method.