Combinator experiments
It seems that having purely immutable basic block
graph is quite impractical as any changes requires
recreation of whole graph. This is especially
complicated due to cycles.

As a simpler solution we just make terminator of
the block to be mutable and implement combinators
on blocks as mutations of the graph.