New c-style vararg interop
Previous scheme with having Any as a type did not work as we would
loose the precise type of the arguments that is necessary to undo
the boxing damage done by erasure.

Now we have a magnet-like Vararg type that is used to insert
conversions from T to Vararg with corresponding classtag information
attached.

Fixes #95, #33