Quick and dirty mean to show callee type and method name
Before each virtual method call, the actual type of the receiver is
shown, along with the name of the method that is called.

The next step is to gather all that information to be able to determine
what is the actual behavior of the program at runtime, to determine
which call sites can be optimized.

Note: Currently, we will get a lot of clashes for calls that could
actually be optimized. For instance, if many types define a method named
`foo`, where all are effectively final, we would still think that the
method foo must go through dynamic dispatch, because we'll think that
the method is declared in a common supertype. Should we distinguish by
call site (use the instruction number maybe?)