Implicit basic blocks
This commit introduces major simplification in terms of how we treat NIR
in-memory. It changes two-level structure of code(sequence of blocks
that are themselves sequences of instructions) into a single-level one.
Block labels are just another kind of instruction now.

Main goal of this refactoring is to simplify transformation passes that
need to change control-flow of the program. Accidentally this also
simplifies front-end code generation a bit.