Execute static method call before merging
Before this commit, inline caching created different branches to
retrieve the pointer to the right method to call, and then merged to a
new block where the call was executed.

Unfortunately, this prevents LLVM from performing some optimizations and
degrades the performance of the program.

We now move the call to the same branch where the method pointer is
retrieved and merge after executing the call.