Introduce infrastracture to test compiler & optimizer
The testing compiler is made of two distinct components, as is the
scala-native compiler:
 - The testing NIRCompiler
 - The testing Optimizer

First, the testing NIRCompiler is used to compile scala code to NIR and
test the `nscplugin` project.

Second, the testing Optimizer is used to test the optimizer, and
ensuring that it performs the right transformations during its passes.

There are now 3 environment variables that contain all the info the
testing compiler needs:

 - scalanative.testingcompiler.cp
   It contains everything the testing is going to need. That classpath
   is used to create a new classloader from which the testing compiler
   will be loaded.
 - scalanative.nscplugin.jar
   The absolute path to the packaged nscplugin.
 - scalanative.nativeruntime.cp
   Everything that the linker needs to have on classpath to do its job.