Initial java.util.regex support with posix regex
Cleaned up, following the style (but not location) of scala-js for regex.
Probably will move to the (correct?) place in the next commit, and also finish some of the stubs.
When all stubs do something (and have some tests), flags should come but it is probably better to skip
adding the flag logic using regex.h and move directly to using the PCRE library directly by then, adding flags to its calls.

Tried to change the interface to be java.util.regex, but can’t get it to work (errors with cannot link: `@java.util.regex.Pattern$::matches_class.java.lang.String_trait.java.lang.CharSequence_bool`). Added some tests for the PosixRegex, but for some reason it throws `java.lang.RuntimeException: Nonzero exit code: 134`. This is with 0 asserts or 3 asserts. Never gets to the third assert. May be a memory issue

Dirty hack while I look up how to hook the real structure (or jump directly to PCRE) to avoid buffer overflows. Moved everything into java.util.regex and changed the interface so is compatible with java, changed the few tests I added to work with that, too