Add support for runtime tags
Runtime tags is the way how we smuggle types across the erasure border.
Previously we used ClassTags for the same purpuose, the only problem
with those is that they don't scale that well to capture all the
necessary information about generic types we care about (e.g. pointers,
tuples etc.)