Box and unbox results of function pointers
When using the values of function pointers boxing (or lack of it) introduced
by the Scala compiler interacts in a  weird way with the nir code gen phase.
For proper boxing instructions, integer and float types of values now have to carry
around the information about the initial Scala type (to distinguish between
the integer values that have the same bit-length).