Prevent random stack overflow errors
I did investigation and I see that no more difference between pictures
after 30+ depth, but as it a recursive function we eventually will get
stack overflow (I think it's possible to make it @tailrec but it's not
at the moment). I added limit for 100 depth (bounces for a ray) to make
demo stable without any visual loss.