Test interface (#755)
* Make `StackTraceElement` closer to Java's impl

* Add initial version of the test interface

There are still many rough edges and issues with it. See comments in the
code for the different issues to tackle :)

* Teach Travis how to build scalacheck

* scalafmt

* Cleanup

* Remove debug info from serializers

* Fix how tests are instantiated

We were assuming that tests are always inside objects, but that's not
true. Fortunately, fingerprints tell us all we need to instantiate the
tests.

* Don't do `rebuild` twice in Travis CI

* scalafmt

* Correctly convert variants of `struct sockaddr`

Previously we would just copy a pre-defined number of bytes, now matter
what the actual underlying struct was. Now, we use the length of the
input struct to determine its type, and correctly convert to the
operating system implementation, so that no information is lost.

* %s/PreLoaded/Preloaded/g

* Remove `stubs` project

* Remove impls in `Class.getField` and `Field.get`

* Deprecate `JSExportDescendent{Classes,Objects}`

* Simpler serialization

* Add implicit classes for `sockaddr` and friends

* Give names to fields of structs for sockets

* Fix review comments

* Prefix test name with `_root_.`

* Add NativeFramework to run our unit tests

* Remove utest and scalacheck tests

* Remove old, dead code