Fix #863: start the test interface server on the sbt side (#874)
Previously, sbt would tell the distant program to start a server on a
given port. Unfortunately, this would prevent tests from running if the
randomly allocated port was already used, and couldn't recover from this
situation.

This commit moves the server part to the sbt plugin. Given a complete
implementation of `java.net`, it's easier to select a free port and
make sure that we can listen on this port. Once we're sure everything is
fine, we can start the distant program and instruct it to connect to the
right port.