Support manipulating JAR files with java.util.jar
This commit adds support for almost all the classes defined in
`java.util.jar` (`JarInputStream` and `JarOutputStream` are missing and
will be added in a subsequent commit.)

Note that this implementation doesn't support verifying JARs signatures.
Everything is already supported on the `java.util.jar` side, but the
classes in `java.security` are only stubs. Implementing these classes
should be sufficient to verify JAR signatures.

This code has been ported from Apache Harmony.