Annotation to mark stubs
This commit adds an annotation `@stub` that can be added to method
definitions. Methods that have the `@stub` annotation will not link by
default (i.e. if something depends on it, it won't be resolved).

To allow `@stub` methods to be linked, one needs to set `linkStubs =
true` in `scala.scalanative.tools.Config`.

Fixes #971