Fix #867: Support `scala.util.Properties`
`scala.util.Properties` works by reading the resource
`/library.properties` using `getResourceAsStream`, which we don't
support. To be able to use `scala.util.Properties`, we rewrite the lhs
of `lazy val scalaProps` (where the resource is read), so that we return
a pre-populated `java.util.Properties` instance if the resource to read
is `/library.properties`. The original implementation is used otherwise.

The content of `scala.util.Properties` that is returned is the same as
the one that was available when the `scalalib` was compiled.

Fixes #867