Avoid virtual dispatch in default == and ## (#990)
Previously, we introduced a peephole optimization to short-circuit the
lookup of scala_== and scala_## to point to the overriden equals and
hashCode directly.

The default implementation, on the other hand still has had a problem of
performing double virtual dispatch. This change addresses the problem by
manually inlining the identity-based equality and hash code.

Additionally we also inline hashCode for longs to avoid box allocation.