Fix #979: Caching of stack trace elements on MacOS
To avoid re-computing all stack trace elements for every exception, we
cache the stack traces. Previously, we were using the address of the
first instruction of a function as a key to retrieve the corresponding
stack trace element. It turns out, however, that retrieving this
information doesn't work as expected on MacOS, which was the cause for
issue #979.

Fortunately, we can also use the current instruction pointer inside each
stack frame as key to get the same effect. This information can be
queried using libunwind and gives consistent results on MacOS and on
Linux.

Fixes #979