Add `name` to `Mode` and `default`
The `default` method in `object Mode` describes the default value for
the `Mode`. It is currently set to `Debug`, following what was
previously done.

`Mode` instances also get a new attribute `name` which is a string
representation of the `Mode`.

Finally, add `apply` method to `Mode` to get an instance from a String.