Fix TimeSuite test failures (#1167)
The TimeSuite tests was failing for me on both OSX and Linux. I
discovered three issues:

1) The difftime stub had the wrong result type
2) I'm not exactly clear on what the
     `localtime() should transform day zero...`
   test was trying to do, but it did not seem to correctly use the
   semantics of localtime and ascstring
3) The stack allocations would cause the test suite to frequently crash
   for me unless I wrapped the tests in a try { } catch { }

The first issue was easy to fix by changing the type signature. I
re-interpreted the second test and changed it to so that it should
adjust the epoch time by the local time zone offset to that the
ascstring returns "Sun Jan  1 00:00:00 1900\n" no matter how the system
is configured. The third issue was fixed by using zone allocation
instead of stack allocation.

I also deleted the unused test.

Fixes: #1167

Bonus: add daylight, tzname and tzset methods