Fix #746: Add support for java.lang.{Process, ProcessBuilder} (#1082)
* Alphabetize time methods

No functional difference.

* Fix TimeSuite test failures (#1167)

The TimeSuite tests was failing for me on both OSX and Linux. I
discovered three issues:

1) The difftime stub had the wrong result type
2) I'm not exactly clear on what the
     `localtime() should transform day zero...`
   test was trying to do, but it did not seem to correctly use the
   semantics of localtime and ascstring
3) The stack allocations would cause the test suite to frequently crash
   for me unless I wrapped the tests in a try { } catch { }

The first issue was easy to fix by changing the type signature. I
re-interpreted the second test and changed it to so that it should
adjust the epoch time by the local time zone offset to that the
ascstring returns "Sun Jan  1 00:00:00 1900\n" no matter how the system
is configured. The third issue was fixed by using zone allocation
instead of stack allocation.

I also deleted the unused test.

Fixes: #1167

Bonus: add daylight, tzname and tzset methods

* Add parens to getFD

This makes it consistent with the getFD method of FileOutputStream (and
the java api).

* Don't alias protected InputStream

Reusing the in variable name makes it difficult to mutate the protected
`in` variable in FilterInputStream.

* Whitespace: alphabetize unistd.scala

No functional change.

* Add user.dir system property

Since this property was missing, I tried to implement it in the same way
that java does. The test that I add essentially tests that both java and
scala native return the same "user.dir" value since it relies on the sbt
System property.

* Add pthread files

These were copied from #1068
(https://github.com/scala-native/scala-native/pull/1068)
since there was overlap between the two prs.

They are just boilerplate anyway.

* Implement #746 Add support for java.lang.{Process, ProcessBuilder}

This commit adds support for the java Process api for posix platforms.

Notes:

* The lack of thread support made this challenging. The naive approach
   to reaping the processes is to add a monitor thread that blocks on
   waitpid and then updates the process. Because scala native doesn't
   support threads yet, this must be done with a thread in native code.
   I initially tried adding minimal pthread apis to scala native, but
   invariably the pthread would allocate memory and the garbage
   collector would eventually crash. To get around this, I wrote the
   monitor thread in c.

   Because the main thread needs to update the process, the monitor
   thread must interrupt it. I ran into problems when the main thread
   was interrupted while reading from the piped file descriptor of the
   child process. To get around this, I add a global mutex that is
   shared by the process and the monitor thread. Most of the process
   methods are guarded by this mutex and the monitor must acquire it
   before it raises the SIGUSR1.

   It should be straightforward to port the monitor thread to scala and
   use java synchronized/wait/notify instead of the mutex/cond approach
   used here once scala native threads are ready.

* To avoid file descriptor leaks, it's a good idea to drain the piped
   file descriptors after the process terminates. We store the drained
   data (if any) into a ByteArrayInputStream which should allow the
   bytes to be freed when the Process is garbage collected if the user
   doesn't explicitly close the input stream. If we didn't do this, we
   wouldn't need to interrupt the process and instead could just
   communicate the process exit value through a heap allocated pointer.

I initially forked from @hwielenberg* to get started, but ultimately
rewrote pretty much all of the code. I'm sure some of the lines in
ProcessBuilder.scala haven't changed much though and I wanted to be sure
to acknowledge that initial work.

*https://github.com/hwielenberg/scala-native/tree/scala-native-746