Fix stack corruption in release
At some point I switched from fork to vfork and while tests passed, I
never tested in release mode. I discovered while debugging the
ProcessSuite hangs that invoking a process always segfaults in release
mode on my mac. I believe this was because the child process modifies
the stack of the parent. The fix is to just create a new function that
runs the child code so that the child creates a new stack frame.

After this change, I ran a bash script that invoked a simple scala
native program that called git status and was compiled in release mode.
In over 1000 iterations of the script, I saw no crashes. (Prior to the
change, it always segfaulted)