Move native/time.scala to posix/time.scala
The time.scala file that added the methods in time.h was inconsistently
located in the native package. This commit moves it into the posix
package and also unifies the declaration of the time types. The time_t
and timespec types are declared in time.h, while timeval is declared in
sys/time.h. I removed the redundant declarations.