Fix undefined process behavior
It turns out that it's unsafe to signal back onto the main thread if the
main thread is performing gc. To work around this, I made it so that the
process results are cached in native code and the scala native process
must fetch them either by explicitly calling waitFor or by implicitly
calling any of a number of relevant apis. In particular, whenever the
user calls a function that checks the exit value, we check the native
cache and whenever we read or check the available bytes.

Unlike the old implementation, we will leak native file descriptors (and
an int -> int map entry) if the user never calls waitFor (or one of the
other functions that can have the implicit side effect of draining the input
streams and setting the proc result). I think this can be fixed by
caching the output and error streams in the native code as well, but I'm
going to defer that work.