Move platform specific code into platform.c
We want to move away from having portable code implemented with runtime
branching, so I moved the setting of the os.name (and version on linux)
into platform.c.

I also added the java.io.tmpdir system property on all platforms. The
semantics are different than the jvm on osx (which uses a user
specific path that can only be accessed using objective c apis in the
foundation framework). Instead of the user specific temp directory, it
just uses /tmp, which is probably fine for most applications. It also
isn't possible to precisely determine the os version on osx without
shelling out to the `sw_vers` command or using different objective c
apis from the foundation framework. For now, we'll just leave the os
version blank on osx (and windows -- there are different issues with
getting this to work correctly on windows and right now there seem to
be very few windows users of scala native).