Set os.name system property
I noticed that scala.util.Properties.isMac didn't work because the
"os.name" system property wasn't set. I added this functionality for OSX
and Linux, but not windows because I don't know what it's supposed to
run. I only set the version for Linux because uname can't reliably used
to get the os.version string. I add support for that in a subsequent
commit, but it requires adding objective c support to the toolchain,
which I imagine may be contentious.

Bonus: move up initializing of streams so that println may be used in
loadProperties