Update createTemp{Directory, File} semantics
The java.nio.Files.createTemp{Directory, File} apis do not require a
minimum length for the file prefix the way the
java.io.File.createTempFile api does. In addition, in the jdk, the temp
directory is by default taken from the java.io.tmpdir system property,
which is not /tmp in osx.