Optimize getName
It is quite slow to repeatedly all split on the underlying path string.
To fix this, we can cache the offsets into the parts of the string and
refer to the offset array whenever we need a part. An alternative would
be to just cache the parts directly, but I think that has higher memory
utilization so most jvm implementations use a lazy list of offsets.