nio.Files -- Improve copy and move semantics
I noticed that Files.copy didn't do the right thing for directories.
They were treated as files, which caused an exception when trying to
read from the InputStream of the source. The java specification is that
copy should create an empty directory if it doesn't exist. If the target
is an empty directory, then its a no-op.

Files.move should replace the existing file or directory as long as the
target doesn't exist, or the REPLACE_EXISTING option is set.

I also added support for copying the file attributes.