Files -- optimize list and walk
I noticed that Files.list and Files.walk were much slower in scala
native than on the jvm. I optimized the Files._list method to narrow the
gap. Instead of delegating to File.list (which wraps an array), I add a
common helper that returns an Array[T] where T is created by applying
a function f, (String, Short) => T to the dirent name and type. By
including the file type in T, we can avoid an extra call to stat, which
is expensive.

There is no more need to call isDirectory in list or walk because
opendir will return -1 and set errno. This will get turned into an
appropriate UnixException depending on whether the path name doesn't
exist or is not a directory.