Support reading and writing more than 1 byte
File IO is slow in scala native because all of the read and write
options go one byte at a time. This commit updates FileChannelImpl to
make bulk operations on the input byte buffers. I saw an O(10x)
performance improvement in a few of the ammonite ops tests (total
runtime for one test decreased from 10 seconds to 1 second. This teat
read from ~150 source files and wrote the concatenated results to an
output file).