Update semantics of NativePosixFileAttributeView
On the jvm, the attributes method of NativePosixFileAttributeView
stats the file and returns all of the file attributes in one shot. It
doesn't, however, memoize the readAttributes method, which allows the
caller to get updated attributes from the same view. In this commit, I
rework things so that attributes is a method that calls stat and
memoizes the result.

This improves performance because stat is relatively expensive.