Provide the timeOps methods to access timeval CStruct2 fields by name.
This is the second of four (4) pull requests (PRs) submitted to fix issue
#1128, "ServerSocket and Socket do not work in nativeMode set to release".

This PR does not directly address the problem in #1228. Rather it
provides infrastructure to make the eventual fix easier/simpler.

A big thank you to the author of select.scala for the tv_*
manipulation routines in selectOps.

I discovered these exceedingly useful methods whilst working on the
eventual fix for #1128. It seemed to me that  methods which manipulate
CStruct timeval fields should reside in sys/time.scala, close to the
datatype they manipulate.

So, I committed the sin of code duplication and copied, mutatis mutandis,
the tv_* methods to timeOps in sys/time.scala. Authors of new code can
now more easily discover and use these methods.

To reduce duplicate code, it makes sense to remove the tv_* methods
from selectOps.  That would be a breaking change for existing code
using those methods.  I hope to submit a PR targeted to just that
change for consideration for the next release which allows breaking
changes (0.4.0?).

This change should work on both 64 & 32 architectures.

Documentation:

	sys/time.scala is currently not documented in
        docs/lib/posixlib.rst. The addition of timeOps would not,
	by itself, trigger a documentation change. The
	the underlying absence should probably be fixed by
	the next person to touch sys/time.scala.

        The first PR in this series, #1242 poll.scala, requires
	a change to posixlib.rst. My plan is to bundle both changes
	to the one file into a later pull request in this series.

        That avoids delaying submitting this PR until after PR #1242
	#has been accepted.

Testing:
	 Tested on both 64-bit & X86 32-bit.

	 java.net SocketSuite & ServerSocketSuite have SO_TIMEIUT
	 tests which, eventually, call these routines to get and set
	 timeval fields.