Add posix poll & six others to the list of implemented files.
This is the third of four (4) pull requests (PRs) submitted to fix issue
#1128, "ServerSocket and Socket do not work in nativeMode set to release".

This PR does not directly address the problem. It documents/announces
posix.scala, which was created in an earlier PR in this series, intended
to make the eventual solution easier.

This PR should be accepted after PR #1242.

As long as I was in the file, I did a bit of "Person Scout"
make-the-world-better by adding six files which now exist in
posix or posix.sys but which were not documented.

I will create an issue for ioctl.scala. It is not a proper posix
declaration. More detail in the soon to be created issue.

Documentation:

        In PR #1242 I promised to document poll.scala. This completes
	that promise.

Testing:
	 Tested by manually building docs/lib, examining the
	 html output, and testing the links to the posix documentation.
	 Since PR #1242 has not been accepted yet, that link is currently
	 broken. That should resolve once PR #1242 is accepted.