This is the fourth and last of four (4) pull requests (PRs) submitted to fix issue #1128, "ServerSocket and Socket do not work in nativeMode set to release".
This PR directly address the problem in #1128. It is a candidate for
the next release (0.4.0?) which accepts breaking changes.
This PR relies upon and provides files submitted in earlier PRs #1242
(poll) & #1243 (timeOps)

This PR avoids using select.scala with its still present memory
difficulties. It uses the more efficient and easier to get correct recently
implemented poll.scala.

I simplified the code flow, so I could convince myself that it was
reasonably correct. The compiler should inline & optimize the functions
I added to better comply with DRY (do not repeat yourself). Even if it is
not that mature, the execution time of a few functions is trivial compared
to a going to the net device and vanishingly small compared to a trip over
the wire to a remote host.

To improve my confidence in correctness, I added error checking and errno
reporting. System calls, _especially_ network calls, do not always succeed.
Knowing that the call failed and the reason (errno) is _way_ better than a
generic "Can not connect" message. I also removed contractions.

These last changes alter the strings which might be seen by an end user.
They will also break any developer forced to parse output strings (yeech!).
As a result, this PR should be considered breaking.

These changes should all work on both 64 & 32 architectures.

Documentation:

	The release notes for the scala-native version which contains
	this change should mention that #1128 is fixed. It should
	also mention additional error detection and reporting has been
	added. Some Exception strings have been changed to remove
	contractions and report errno.

Testing:
	 Tested on both 64-bit & X86 32-bit.

	 The tests below exercise PlainSocketImpl. All pass, no
	 failures are introduced anywhere else.

	     unit-tests: java.net.SocketSuite, java.net.ServerSocketSuite
	     scripted-tests: java-net-socket, java-net-server-socket