Populate posix sys/types.scala on the road to a single point of baseline posix type declarations. This allows more closely following the posix documentation whilst consolidating Cmumble type declarations. Get all/many uses of a posix type right or wrong at the same time. No, or at least minimal, hard to find and probably wrong solo type declarations.
PR #1248 fixed issue #1128 by using poll() and punting fixing select().
A proper fix to select() will be  less complicated with at least
the time* types populated in sys/types.h.

This is a "reduction in strength" fix, rather than a total fix. It
provides a foundation upon which later changes can build.

It is easy to change posix time.scala and sys/time.scala to use the
new foundational type definitions in sys/types.scala. They are
included in this PR.

Changes to stat.scala and unistd.scala are deferred to future PR.
Changing unistd.scala is easy enough, but stat.scala relies upon
some type definitions in unistd.h. off_t is the major concern,
particularly on 32 bit systems.  stat.scala and unistd.scala, as is,
are correct on 64 bit. It is just that the software engineering leaves
something to be desired.

On 32 bit systems, off_t in unistd.scala triggers a discussion of how
small & large file support is to be added (or documented as missing).
The current definition is almost certainly wrong; should be CInt.
Plenty of work for later.

I could not find unit-tests for stat(), and am reluctant to touch
stat.h before I can create such. Call it cowardice or wisdom.

64/32 bit issues:

      None known with this PR.

Documentation:

      No user facing documentation changes once PR #1244 is accepted.
      That pull request announced posix types.scala.

Testing:

 	Tested on both 64-bit & X86 32-bit. Both runs passed.