Decouple s.c.i.Range from Java formatting (#1269)
* Decouple s.c.i.Range from formatters

Here we reimplement Range::description method to use
string builder rather than formatter. This halfs number of
reachable methods on an empty project (from 9k to 4.5k).
This naturally brings down the binary for hello world project
(from 4m to 1.9m) and total linking time.

* Fix equals/hashCode short-circuiting

Those methods are not guaranteed to be virtual.