Improve compatibility and performance of File I/O (#1216)
* Add UnixException and AccessDeniedException

This commit adds the UnixException object which can be used to generate
exceptions when a native posix file system api call fails. I also added
the AccessDeniedException which was thrown by UnixFileSystemProvider but
not actually implemented anywhere.

* Update createTemp{Directory, File} semantics

The java.nio.Files.createTemp{Directory, File} apis do not require a
minimum length for the file prefix the way the
java.io.File.createTempFile api does. In addition, in the jdk, the temp
directory is by default taken from the java.io.tmpdir system property,
which is not /tmp in osx.

* Optimize removeRedundantSlashes

It is often the case that the path string contains no redundant slashes.
When this is true, we still always create a new StringBuffer and result
string when we could just return the string itself (or the substring
that removes the trailing slash). On my machine, this reduced the
average runtime of this method from O(1us) to O(300ns). The extra
complexity is minimal so the optimization feels worth it.

I also added some test cases because in pursuing alternative
optimizations, I found some edge cases what weren't tested.

* Optimize getName

It is quite slow to repeatedly all split on the underlying path string.
To fix this, we can cache the offsets into the parts of the string and
refer to the offset array whenever we need a part. An alternative would
be to just cache the parts directly, but I think that has higher memory
utilization so most jvm implementations use a lazy list of offsets.

* Move UnixPath to jvm semantics

On the jvm, the redundant slashes are removed for methods like toString,
equals and hashCode. I removed the private val from rawPath so that the
reference isn't retained after class construction.

* Files.readAllBytes -- read directly

I noticed that readAllBytes was much slower than on the jvm. Prior to
this change, small files took O(1ms) to read all of the bytes. After the
change, it is more like (30us).

* Files -- optimize list and walk

I noticed that Files.list and Files.walk were much slower in scala
native than on the jvm. I optimized the Files._list method to narrow the
gap. Instead of delegating to File.list (which wraps an array), I add a
common helper that returns an Array[T] where T is created by applying
a function f, (String, Short) => T to the dirent name and type. By
including the file type in T, we can avoid an extra call to stat, which
is expensive.

There is no more need to call isDirectory in list or walk because
opendir will return -1 and set errno. This will get turned into an
appropriate UnixException depending on whether the path name doesn't
exist or is not a directory.

* nio.Files -- Improve copy and move semantics

I noticed that Files.copy didn't do the right thing for directories.
They were treated as files, which caused an exception when trying to
read from the InputStream of the source. The java specification is that
copy should create an empty directory if it doesn't exist. If the target
is an empty directory, then its a no-op.

Files.move should replace the existing file or directory as long as the
target doesn't exist, or the REPLACE_EXISTING option is set.

I also added support for copying the file attributes.

* Support reading and writing more than 1 byte

File IO is slow in scala native because all of the read and write
options go one byte at a time. This commit updates FileChannelImpl to
make bulk operations on the input byte buffers. I saw an O(10x)
performance improvement in a few of the ammonite ops tests (total
runtime for one test decreased from 10 seconds to 1 second. This teat
read from ~150 source files and wrote the concatenated results to an
output file).

* Update semantics of NativePosixFileAttributeView

On the jvm, the attributes method of NativePosixFileAttributeView
stats the file and returns all of the file attributes in one shot. It
doesn't, however, memoize the readAttributes method, which allows the
caller to get updated attributes from the same view. In this commit, I
rework things so that attributes is a method that calls stat and
memoizes the result.

This improves performance because stat is relatively expensive.