Add optimized split (#1220)
It is very slow to have to compile a regex pattern for every call to
split so I provided an optimized version of split that is similar to the
versions provided by other jvm providers. In the case of a happy path
where the separator is a single character that isn't a regex special
character or when the separator is an escaped regex special character,
we use the optimized version.

I made a crude benchmark where I split the string "abcdefghijklnop" * 10
on the letter j 10000 times and the runtime for the fast version was 4us
per iteration while using Pattern.compile.split took 300us.