Sys props (#1218)
* Set os.name system property

I noticed that scala.util.Properties.isMac didn't work because the
"os.name" system property wasn't set. I added this functionality for OSX
and Linux, but not windows because I don't know what it's supposed to
run. I only set the version for Linux because uname can't reliably used
to get the os.version string. I add support for that in a subsequent
commit, but it requires adding objective c support to the toolchain,
which I imagine may be contentious.

Bonus: move up initializing of streams so that println may be used in
loadProperties

* Remove unused private method in java.io.File

This was left behind in 40dd654fe393507f0b1599a450d8c0c1b75d45ec.

* Move platform specific code into platform.c

We want to move away from having portable code implemented with runtime
branching, so I moved the setting of the os.name (and version on linux)
into platform.c.

I also added the java.io.tmpdir system property on all platforms. The
semantics are different than the jvm on osx (which uses a user
specific path that can only be accessed using objective c apis in the
foundation framework). Instead of the user specific temp directory, it
just uses /tmp, which is probably fine for most applications. It also
isn't possible to precisely determine the os version on osx without
shelling out to the `sw_vers` command or using different objective c
apis from the foundation framework. For now, we'll just leave the os
version blank on osx (and windows -- there are different issues with
getting this to work correctly on windows and right now there seem to
be very few windows users of scala native).