Improve codegen scalability
This change improves work balancing for the
codegen and native compilation steps. Previously
we would emit one file per package. Some packages
contain more definitions than the others, which
causes work skew (a few threads do most of the work)
which in turn leads to poor scalability to 6+ cores.

Instead, the work should be split randomly across a
number of files proportional to the number of available
processors. This minimizes the risk of any single file
taking longer to compile than the others.

This change shaves off 1s of compilation on the tests
project on 6 core machine with hyper threading.