Add dedicated instructions for field load/store
Previously we represented field operations through a mix of `Op.Field`
and `Op.Load`/`Op.Store`. The semantics of `Op.Field` was to return an
inner pointer that points directly to the field within a class, and then
subsequent load and stores would perform an operation through that
pointer. This encoding turned out to be tricky to optimize as a
high-level operation in Interflow (#1278) because we also use load and
store for C's unsafe memory operations.

This pr introduces two dedicated instructions `Op.Fieldload` and
`Op.Fieldstore`. Unlike the previous encoding these are only used for
high-level Scala and are never emitted for C-style memory accesses. This
makes it easier to optimize the away without doing any additional
analysis.