Avoid exceptions in stack trace unmangling
Previously we used higher-level Scala for symbol unmangling.
While this code worked perfectly for symbols that are
correctly recognized as Scala-mangled names, it would throw
and then later catch an exception to handle failures. This
is problematic because unmangler already runs on the exceptional
path and throwing an exception while an exception is being
constructed is a good recipe for stack overflow.

Instead, this PR downgrade the code to only use pointers and
manual bounds checks in combination with return codes to handle
errorneous cases. This ensures that no exceptions are being thrown
during unmangling.