Checker: initial infrastructure (#1408)
* Checker: initial infrastructure

This PR introduces initial infrastructure for NIR checking. It checks that all instructions and jumps within methods are well-typed with respect to our erased type system before and after the optimizer. We'll probably extend it with more checks over time but this is the minimal subset which is already quite useful.

The checker is only going to be used for the development of the toolchain itself, we'll keep it off for end-user projects (likely under a flag).

* Remove U{Byte, Short, Int, Long} types from NIR

* Impl and lifted classes are not module classes

* Better lubs based on lineralization

* Drop block param reduction

* Narrow down types at call site based on duplicate signature

* Make sure abstract classes are never allocated

* Drop redundant check

* Fix comment typos

* Test subtyping relation

* Refactor build instrumentation for checker