Add support for first-class values of native.Tag
Currently tags are special because they must be interpreted
at compile-time by the scalac plugin and it is undefined otherwise.
This change introduces first-class value equivalent for tags so
that operations that depend on tags can provide a generic implementation
based on it (this includes sizeof, typeof, pointer operations etc).