Fix regression due to new implemetation of sizeof
In #1424, we unintrinsified sizeof and typeof,
now they are implemented as regular library code.
This regressed some benchmarks due to them not
being perfectly optimized away in some cases.
To fix this we avoid using sizeof in the Arrays
implementation and change typeof to not use ClassTag
but rather use java.lang.Class directly.