Interflow: combine instructions  (#1444)
This PR revives InstCombine optimization, only as part of the Interflow's main optimization pass rather than a stand-alone pass. The goal of InstCombine is to combine pairs or triples of connected instructions and reduce them down to a simpler alternative based on algebraic equivalences. The implementation is based on the delayed instructions introduced in #1437.