Remove now obsolete comment of many lines, wrap one comment line
Now that PR #1436 has been accepted, this PR should now pass CI.
This commit should trigger a rebuild with current mainline.

A lot has changed since the original commit of this PR, particularly
with respect to Locale.  The block comment in the prior version of
FormatterSuite.scala for this PR no longer makes sense, so I removed it.
Better silence than misdirection.

I also wrapped a comment line in FormatterUSSuite.scala for better
readability.  Not enough for a commit by itself, but as long as
the bus was leaving anyway...