Interflow: bring back inlining to debug mode (#1505)
This PR brings back support for some limited inlining in debug mode. It includes:

1. Field accessors.
2. Compiler-generated bridges.
3. Constructors.
4. Methods annotated with `@alwaysinline`.

This is necessary to keep runtime performance of #1426 within realms of reason 
in debug mode. Otherwise we have to pay the cost of boxing on every single pointer operation. 

Due to parallelism in debug mode we have to inline unoptimized version of the 
method, rather than optimized one like in release mode. Given that all of the methods 
which we want to inline here are usually small this shouldn't be an issue.