Interflow: intrinsify getRawType rather than getClass (#1509)
This is a minor refactoring of Interflow's implementation of getClass.

Now, we can avoid doing unsafe load(bitcast[ptr] obj) to get raw type in the polymorphic inlining implementation and just use getRawType instead. Subsequent revisits of the same code can be folded away if the type of the object becomes more specific, unlike unsafe load of the type pointer.

Additionally, implementation of getClass doesn't have to be intrinsified any more because it's just an allocation of a wrapper over raw type.