Bug Hunt. Temporarily disable all re2s unit-test Suites.
The full unit-test Suites work in my local environment both on this PR
branch and on others.  Similarly the re2s Suites in this PR work
in my local environment.

When I run Travis CI, I get errors in j-random unit-tests and
then the continuing, uncessant NullPointerException from the test
runner, because its control socket is closed.  I do not want
to get into a recursive edit fixing the test runner.  Call me either
Chicken or smart.

So factor out the unit-tests from re2s and see if the rest of the
unit-test world in Travis suceeds, as I am expecting.

If I can establish that baseline, I can start enabling tests and
try to identify the slayer.  Agreed, messy to the max.