Fix #1574: emit all class definitions in the same .ll module (#1587)
This a potential fix for #1574. Previously unit tests failed
in debug mode on macOS on machines with 8 virtual cores.
This PR changes the partitioning scheme to bundle all definitions
for a given class together in LLVM .ll files.

I suspect that that the underlying issue could be a bug in LLVM.
It works as expected if we emit a single LLVM module, or use
any of the optimization modes other than -O0 or use lto.