Add a probabilistic fix to "stack safe" test.
This is a micro-edit point fix. A better comprehensive, deterministic fix
is being exercised in work-in-progress for a future PR.

Prior to this commit, the "stack safe" test could intermittently
fail with duplicate groups. This fix increases the size of the string
to drastically reduce the chance of a duplicate name.

Note that this code to generate a group name can still generate, with
a 10 in 62 probability, a digit for the first character of the group
name. Such a name is not a valid Java 8 group name and will cause
the test to fail intermittently and in a hard to debug manner.

Again, a better fix is on the way. Both fixes are in a future PR
to simplify the merge process.