Commit_message.utf-8 --
Lee Tibbert
2020-05-06

During the development of PR #1752 it became evident that StringSuite.scala
was testing the form of String.getBytes() which has been deprecated
since JDK 1.1.  It also became apparent that the recommended form
specifying an encoding was not being tested.

There was an attempt to remove the test of the deprecated form, but that
test was deemed necessary.  In that context, adding the test for
Unicode formats was not appropriate.

This PR adds a test for getBytes() using the extensively used UTF-8
encoding.

No documentation impact.

Manually tested on Java 8.

// -30- //