Fix #1782: Port ScalaSources unpacking from Scala.js (#1799)
hepin1989 and I have worked on this issue together. The Issue
contains most of that dialog.

The port is almost one for one.  There was one small issue
which I documented in the code about the relative location
of the overrides directory or directories: the two projects
differ by one level.

The first attempts to make the change of this PR ran aground
with a missing dependency on nscplugin. I edited nativelib,
which is the bottom dependency, to ensure that the required
nscplugin is built before nativelib is built.  This allows
nativelib and projects which depend upon it to build cleanly
the first time after a root project is created or the first time
after caches are cleaned.

I also moved several `.in(file("foo"))` lines so that the file
has a consistent;