Fix #1754: scripts/gyb.py fails with python3. Merge after PR #1753 (#1757)
The fix should work with all Python 3 installations and Python
2 versions back to Python 2.2, November 2001, also hard days.
I do not know if the rest of the script works with those earlier
versions. I did not add any code to check for minimal Python version.

scripts/gyb.py proved to be innocent. The effective fix was in the file
nativelib/src/main/scala/scala/scalanative/runtime/Arrays.scala.gyb.
The way of iterating over a Python dictionary (Scala Map) changed
from Python 2 to Python 3.  Python 2.7 has code to work with
the Python 3 syntax.

I put .gyb and .scala comments in Arrays.scala.gyb explaining that
the order of output items is sensitive to Python version. I also
suggested that Python >= 3.6, with a defined iteration order, should
be used to run the gyb.py script to avoid non-productive and
time-consuming-to-verify changes in git.

Although not strictly necessary after the above change,
scripts/gyb.py was updated to latest version.
Further information is at the top of that file, where is it
more likely to be found than here.

Comments were added at top of each .gyb file so that generated .scala
files are marked (paraphrase) at the top of the file:
"BEWARE: Do not manually directly edit or your changes will be lost".