Build.sbt: restore nativelib / publishLocal handling (#1804)
This commit reverts a change introduced in
24400e3e78a3ffe9d8c781d4621a7c29a339e6e9,
which unintentionally increases build time.

The purpose of 24400e3e78a3ffe9d8c781d4621a7c29a339e6e99 was to adopt
the proven Scala.jsway of fetching Scala sources in order to speed up
the fetch, especially over slow communications links.

That commit accomplished its main goal. It also revealed a disharmony
in the build.sbt dependency chain when one is using a -SNAPSHOT
version and does a command such as `scalalib / compile` from a
clean, perhaps newly created, repository, with a clean local cache.
The required nscplugin is not found in update, because it is not
there in the described situation. The initial total `rebuild`
recommended in the Contributor's Guide works.