Use reflection instead of source generation in tests (#1802)
Instead of generating sources that collect, register and run the test classes, leverage the newly introduced `@EnableReflectiveInstantiation` annotation for the testing infrastructure.

In particular:
- Load the tests and frameworks via reflection. To that end, annotate `tests.Suite` and `sbt.testing.Framework` with `@EnableReflectiveInstantiation`.
- Implement `org.scalajs.testinterface.TestUtils` in terms of `Reflect` and deprecate it. This will be removed in future versions (probably after 0.4.0). Frameworks should use the reflection API to load test classes.
- Get rid of `PreloadedClassLoader`.
- Get rid of the `NativeTest` sbt configuration.
- Do not generate a test main anymore. Rename `scala.scalanative.testinterface.TestMainBase` to `scala.scalanative.testinterface.TestMain` and use it as the test main class.

A subtle change is in `scala.scalanative.testinterface.ScalaNativeFramework`. The `name` and `fingerprints` are no longer obtained via the `Runner`, but via the enclosed `Framework` object (which is a `NativeFramework`). This is required since we removed the `NativeTest` sbt configuration (it is now basically just the `Test` configuration) and the `Runner` is not available at the moment the above attributes are requested. So, we use the `NativeFramework` from the JVM to obtain them. It is not very elegant, but does the job for the time being. We should refactor this in a less ugly solution, similar to what Scala.js does with its testing adapter. But that deserves its own PR.